Move some graph related code from render/mod.rs to render/graphviz.rs
