+++ src/render/mod.rs
@@ -74,58 +74,9 @@ impl<'a, 'tcx: 'a> Renderer<'a, 'tcx> {
         let rendered_source = source::render_source(self.tcx, frame);
 
         let mir_graph = frame.map(|frame| {
-            let mut mir_graphviz = String::new();
-            graphviz::write(&frame.mir, breakpoints.for_def_id(frame.instance.def_id()), &mut mir_graphviz).unwrap();
-            String::from_utf8(::cgraph::Graph::parse(mir_graphviz).unwrap().render_dot().unwrap()).unwrap()
+            graphviz::render_html(frame, breakpoints.for_def_id(frame.instance.def_id()))
         });
-        let (bb, stmt) = frame.map_or((0, 0), |frame| {
-            let blck = &frame.mir.basic_blocks()[frame.block];
-            use rustc_data_structures::indexed_vec::Idx;
-            (
-                frame.block.index() + 1,
-                if frame.stmt == blck.statements.len() {
-                    if blck.statements.is_empty() {
-                        6
-                    } else {
-                        blck.statements.len() + 7
-                    }
-                } else {
-                    assert!(frame.stmt < blck.statements.len());
-                    frame.stmt + 6
-                },
-            )
-        });
-        let edge_colors = if let Some(ref frame) = frame {
-            let blck = &frame.mir.basic_blocks()[frame.block];
-            let (targets, unwind) = if frame.stmt == blck.statements.len() {
-                use rustc::mir::TerminatorKind::*;
-                match blck.terminator().kind {
-                    Goto { target } => (vec![target], None),
-                    SwitchInt { ref targets, .. } => (targets.to_vec(), None),
-                    Drop { target, unwind, .. } |
-                    DropAndReplace { target, unwind, .. } => (vec![target], unwind),
-                    Call { ref destination, cleanup, .. } => {
-                        if let Some((_, target)) = *destination {
-                            (vec![target], cleanup)
-                        } else {
-                            (vec![], cleanup)
-                        }
-                    }
-                    _ => (vec![], None),
-                }
-            } else {
-                (vec![], None)
-            };
-            format!("let edge_colors = {{{}}};",
-                targets.into_iter().map(|target|(frame.block, target, "green"))
-                    .chain(unwind.into_iter().map(|target|(frame.block, target, "red")))
-                    .map(|(from, to, color)| format!("'bb{}->bb{}':'{}'", from.index(), to.index(), color))
-                    .collect::<Vec<_>>()
-                    .join(",")
-            )
-        } else {
-            "{}".to_string()
-        };
+
         println!("running horrorshow");
         use horrorshow::Raw;
         promise.set(Html(box_html! {
@@ -157,31 +108,8 @@ impl<'a, 'tcx: 'a> Renderer<'a, 'tcx> {
                             p { : message }
                         }
                         div(id="mir") {
-                            style { : Raw(format!("
-                                #node{} > text:nth-child({}) {{
-                                fill: red;
-                                }}
-                                .edge-green > path, .edge-green > polygon, .edge-green > text {{
-                                    fill: green;
-                                    stroke: green;
-                                }}
-                                .edge-red > path, .edge-red > polygon, .edge-red > text {{
-                                    fill: red;
-                                    stroke: red;
-                                }}
-                                .edge > path {{
-                                    fill: none;
-                                }}
-                            ", bb, stmt))}
                             : Raw(mir_graph.unwrap_or("no current function".to_string()))
                         }
-                        script { : Raw(edge_colors + r##"
-                                for(let el of document.querySelectorAll("#mir > svg #graph0 .edge")) {
-                                    let title = el.querySelector("title").textContent;
-                                    if(title in edge_colors) {
-                                        el.classList.add("edge-" + edge_colors[title]);
-                                    }
-                                }"##) }
                     }
                     div(id="right") {
                         div(id="stack") {
